+++ src/components/Tabs/Tabs.tsx
@@ -7,39 +7,41 @@ import PropTypes from 'prop-types';
 export type Children = React.ReactElement<TabPanelProps, typeof TabPanel>;
 
 export interface TabsProps {
-  /** 
+  /**
    * 前缀
    */
   prefix: string;
-  /** 
+  /**
    * 类名
    */
   className?: string;
-  /** 
+  /**
    * 样式
    */
   style?: React.CSSProperties;
-  /** 
+  /**
    * 激活的 tab key
    */
   activeId: string | number;
-  /** 
+  /**
    * 选中的tab改变时
    */
   onChange?: (id: string | number) => void;
-  /** 
-   * tab bar 上额外的元素 
+  /**
+   * tab bar 上额外的元素
    */
   tabBarExtraContent?: React.ReactNode;
   /**
    * children
    */
-  children: Children[]
+  children: Children[];
 }
 
-export interface Nav {
+export interface TabProps {
   id: string | number;
-  tab: React.ReactNode
+  tab: React.ReactNode;
+  className?: string;
+  style?: React.CSSProperties;
 }
 
 /**
@@ -62,22 +64,22 @@ class Tabs extends PureComponent<TabsProps> {
 
   static TabPanel: typeof TabPanel;
 
-  genNavsContents = () => {
+  genTabsContents = () => {
     const { children } = this.props;
-    let navs: Nav[] = [];
+    let tabs: TabProps[] = [];
     let contents: Children[] = [];
-    React.Children.forEach(children, (child) => {
+    React.Children.forEach(children, child => {
       const isElemet = React.isValidElement(child);
       if (!isElemet) {
         return;
       }
       const {
-        props: { tab, id }
+        props: { tab, id, className, style }
       } = child;
-      navs.push({ id: id, tab: tab });
+      tabs.push({ id, tab, className, style });
       contents.push(child);
     });
-    return { navs, contents };
+    return { tabs, contents };
   };
 
   onTabClick = (id: string | number) => {
@@ -85,47 +87,35 @@ class Tabs extends PureComponent<TabsProps> {
     onChange && onChange(id);
   };
 
-  renderNav = (navs: Nav[]) => {
+  renderTabs = (navs: TabProps[]) => {
     const { prefix, activeId } = this.props;
     const classPrefix = `${prefix}-tabs__nav`;
-    return navs.map((element) => {
-      const isElement = React.isValidElement(element.tab);
-      if (isElement) {
-        return (
-          <div
-            key={element.id}
-            onClick={() => {
-              this.onTabClick(element.id);
-            }}
-            className={classNames(`${classPrefix}-link`, {
-              [`${classPrefix}-link--active`]: activeId === element.id
-            })}
-          >
-            {element.tab}
-          </div>
-        );
-      } else {
-        return (
-          <div
-            key={element.id}
-            onClick={() => {
-              this.onTabClick(element.id);
-            }}
-            className={classNames(classPrefix, {
+    return navs.map(element => {
+      return (
+        <div
+          key={element.id}
+          onClick={() => {
+            this.onTabClick(element.id);
+          }}
+          className={classNames(
+            classPrefix,
+            {
               [`${classPrefix}--active`]: activeId === element.id
-            })}
-          >
-            {element.tab}
-          </div>
-        );
-      }
+            },
+            element.className
+          )}
+          style={element.style}
+        >
+          {element.tab}
+        </div>
+      );
     });
   };
 
   renderContent = (contents: Children[]) => {
     const { prefix, activeId } = this.props;
     const classPrefix = `${prefix}-tabs__content`;
-    return contents.map((element) => {
+    return contents.map(element => {
       const { id } = element.props;
       return (
         <div
@@ -142,7 +132,7 @@ class Tabs extends PureComponent<TabsProps> {
 
   render() {
     const { prefix, className, style, tabBarExtraContent } = this.props;
-    const { navs, contents } = this.genNavsContents();
+    const { tabs, contents } = this.genTabsContents();
 
     const classPrefix = `${prefix}-tabs`;
     const classes = classNames(classPrefix, className);
@@ -150,7 +140,7 @@ class Tabs extends PureComponent<TabsProps> {
     return (
       <div className={classes} style={style}>
         <div className={`${classPrefix}__nav-wrapper`}>
-          <div className={`${classPrefix}__navs`}>{this.renderNav(navs)}</div>
+          <div className={`${classPrefix}__navs`}>{this.renderTabs(tabs)}</div>
           {tabBarExtraContent && (
             <div className={`${classPrefix}__extra-content`}>
               {tabBarExtraContent}