+++ src/components/index.ts
@@ -19,3 +19,5 @@ export { default as Table } from './Table';
 export { default as Search } from './Search';
 export { default as Select } from './Select';
 export { default as Radio } from './Radio';
+export { default as Checkbox } from './Checkbox';
+export { default as Menu } from './Menu';